requirements.txt
flask
flask_wtf
flask_sqlalchemy
flask_login
flask_bcrypt
flask_mail
email-validator
wtforms
jwt
pillow